+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 0.0.1
 
 - Initial release
+
+## 0.0.2
+
+- stop loop on window focus lose
+- stop loop when re-running the command while the prev didnt finish